ui_requirements.txt
# Core data processing (minimal)
pandas>=1.5.0
numpy>=1.24.0

# Image processing (basic)
Pillow>=9.5.0

# Web framework
streamlit>=1.42.2
streamlit-geolocation

# HTTP client for API communication
requests>=2.31.0

# Visualization
matplotlib>=3.6.0
plotly>=5.15.0

# Utility libraries
python-dotenv>=1.0.0

# Configuration
pydantic>=2.0.0